更新服务器行为树；增加行为树变量组件并增加示例；

+++ Server/Model/Module/Demo/Config/BehaviorTreeConfig.cs
@@ -1,6 +1,6 @@
 namespace ETModel
 {
-	[Config(AppType.ClientH |  AppType.ClientM | AppType.Gate | AppType.Map)]
+	[Config((int)(AppType.ClientH |  AppType.ClientM | AppType.Gate | AppType.Map))]
 	public partial class BehaviorTreeConfigCategory : ACategory<BehaviorTreeConfig>
 	{
 	}
